+++ lib/catalog/album.rb
