+++ lib/catalog/item.rb
@@ -1,5 +1,3 @@
-require_relative 'label'
-
 class Item
   attr_reader :publish_date, :label, :date, :genre, :author
 
@@ -10,20 +8,22 @@ class Item
   end
 
   def can_be_archived?
+    year = Time.new.year
+    year - @publish_date.year > 10
   end
 
   def move_to_archive
   end
 
   def add_genre
+    @genre = genre
+    genre.items.push(self) unless genre.items.include?(self)
   end
 
   def add_author
   end
 
-  def add_label(label: Label)
-    @label = label
-    label.items << self unless label.items.include?(self)
+  def add_label
   end
 
   private :can_be_archived?
